// src/app/models/product.ts
export class Product {
    Id: number;
    Name: string;
    Description: string;
    StorePrice: number;
    Long: number;
    Wide: number;
    High: number;
    Weight: number;
    FinalPricePesos: number;
    FinalPriceDollars: number;
    StoreId: number;
    CategoryId: number;
    DeliveryTime: number;
    Removed: boolean;
}